creation module pour les calculs
